Add files via upload
Check out this program that I just completed for SVM-based diabetes prediction.

The data set which I have used here is originally from the National Institute of Diabetes and Digestive and Kidney Diseases. The objective of the dataset is to diagnostically predict whether a patient has diabetes based on certain diagnostic measurements included in the dataset.

Information about dataset attributes -
Pregnancies: To express the Number of pregnancies
Glucose: To express the Glucose level in blood
BloodPressure: To express the Blood pressure measurement
SkinThickness: To express the thickness of the skin
Insulin: To express the Insulin level in blood
BMI: To express the Body mass index
DiabetesPedigreeFunction: To express the Diabetes percentage
Age: To express the age of the person
Outcome: The final result ‘1’ is a diabetic patient and ‘0’ is not a diabetic patient

The procedure begins with loading the data set and continues with validations such as data types and nullity. The data is then separated into features and class labels and converted into array.The resulting data is split into train and test data with #train_test_split and scaled with #MinMaxScaler. The simple data thus obtained is used to make prediction using the classifier #KNeighborsClassifier.

Attached is the document of the same

#predictiveanalytics #knn #svm #pythonprogramming #googlecolab